requirements.txt
flask
flask-cors
python-dotenv
requests
pyautogui
pyperclip
psutil
pygetwindow
keyboard
mouse
pillow
python-docx
pypdf
python-pptx
edge-tts
pytesseract
pywin32
